Move Settings related types to DataFeeds sub-namespace
